Caching problems

+++ lib/max_gallery_web/live/import_live.ex
@@ -7,6 +7,11 @@ defmodule MaxGalleryWeb.Live.ImportLive do
 
   def mount(params, %{"auth_key" => key, "user_auth" => user}, socket) do
     group_id = Map.get(params, "page_id")
+    
+    # Calculate current user storage
+    current_size_gb = Utils.user_size(user)
+    max_size_gb = Variables.max_size_user()
+    storage_exceeded = current_size_gb >= max_size_gb
 
     zip? = Map.get(params, "zip")
 
@@ -37,7 +42,10 @@ defmodule MaxGalleryWeb.Live.ImportLive do
         user: user,
         loading: false,
         zip: zip?,
-        page_id: group_id
+        page_id: group_id,
+        current_size_gb: current_size_gb,
+        max_size_gb: max_size_gb,
+        storage_exceeded: storage_exceeded
       )
 
     {:ok, socket, layout: false}
@@ -74,22 +82,32 @@ defmodule MaxGalleryWeb.Live.ImportLive do
     group_id = socket.assigns[:page_id]
     zip? = socket.assigns[:zip]
     user = socket.assigns[:user]
+    storage_exceeded = socket.assigns[:storage_exceeded]
 
-    consume_uploaded_entries(socket, :file_import, fn %{path: path}, %{client_name: name} ->
-      if zip? do
-        if Utils.zip_valid?(path) do
-          Context.unzip_content(path, user, key, group: group_id)
+    if storage_exceeded do
+      # Don't upload if storage limit is exceeded
+      {:noreply, socket}
+    else
+      consume_uploaded_entries(socket, :file_import, fn %{path: path}, %{client_name: name} ->
+        if zip? do
+          if Utils.zip_valid?(path) do
+            Context.unzip_content(path, user, key, group: group_id)
+          else
+            File.rm!(path)
+          end
         else
-          File.rm!(path)
+          case Context.cypher_insert(path, user, key, name: name, group: group_id) do
+            {:ok, _id} -> :ok
+            {:error, "storage_limit_exceeded"} -> :storage_limit_exceeded
+            {:error, _reason} -> :error
+          end
         end
-      else
-        Context.cypher_insert(path, user, key, name: name, group: group_id)
-      end
 
-      {:ok, nil}
-    end)
+        {:ok, nil}
+      end)
 
-    {:noreply, push_navigate(socket, to: "/user/data/#{group_id}")}
+      {:noreply, push_navigate(socket, to: "/user/data/#{group_id}")}
+    end
   end
 
   def handle_event("cancel", _params, socket) do
